Fix `elimDeadDsts` peephole optimization in native codegens
The `elimDeadDsts` peephole optimization attempts to eliminate
instructions whose destinations are dead.  There was limited code for
preserving instructions that are immediately followed by a `setcc`
instruction or a `if` transfer.  However, it did not preserve
instructions that were immediately followed by an `adc` or `sbb`
instruction.  This would lead to mis-compilation of
`Word64_{add,neg,sub}CheckP` primitives.  For example, a
`Word64_addCheckP` could initially be translated to:

    movl src1Lo,tmpLo
    movl src1Hi,tmpHi
    addl src2Lo,tmpLo
    adcl src2Hi,tmpHi
    seto dst

The "tmpLo" destination in the "addl src2Lo,tmpLo" instruction is dead
and so the instruction would be eliminated; in turn, the "tmpLo"
destination in the "movl src1Lo,tmpLo" instruction is dead and so the
instruction would be eliminated.  However, although the "tmpHi"
destination in the "adcl src2Hi,tmpHi" instruction is dead, the
instruction is preserved because of the subsequent "seto"
instruction.  The end result would be:

    movl src1Hi,tmpHi
    adcl src2Hi,tmpHi
    seto dst

which fails to set "dst" appropriately.

Now, in addition to preserving instructions that are immediately
followed by a `setcc` instruction or an `if` transfer, an instruction
immediately preserved by a `adc` or `sbb` instruction is preserved.