Fix x86-codegen translation of `Word64_{add,neg,sub}CheckP` prims
Previously, a `Word64_negCheckP` was initially translated to:

    movl $0,tmp
    subl srcLo,tmp
    movl $0,tmp
    sbbl srcHi,tmp
    seto dst

With commits 97315a4 and 3a0a42d, the "subl srcLo,tmp" instruction is
not eliminated by the `elimDeadDsts` peephole optimization.  However,
register allocation unconditionally rewrites `movl $0,dst` as
`xorl dst,dst`, leaving code like:

    xorl tmp,tmp
    subl srcLo,tmp
    xorl tmp,tmp
    sbbl srcHi,tmp

However, while a `mov` preserves the condition code flags, a `xorl`
instruction clears the CF flag, which leads to incorrect flags being
used and set by the `sbb` instruction.

Similarly, a `Word64_addCheckP` was initially translated to:

    movl src1Lo,tmp
    addl src2Lo,tmp
    movl src1Hi,tmp
    adcl src2Hi,tmp
    seto dst

which, if src1Hi were 0, would result in

    movl src1Lo,tmp
    addl src2Lo,tmp
    xorl tmp,tmp
    adcl src2Hi,tmp

which leads to incorrect flags being used and set by the `adc` instruction.

The existing 64-bit primitives avoid this problem by performing all of the `mov`
instructions before the arithmetic instruction.

Now, a `Word64_negCheckP` primitive is initially translated to:

    movl $0,tmpLo
    movl $0,tmpHi
    subl srcLo,tmpLo
    sbbl srcHi,tmpHi
    seto dst

which can be correctly rewritten to:

    xorl tmpLo,tmpLo
    xorl tmpHi,tmpHi
    subl srcLo,tmpLo
    sbbl srcHi,tmpHi
    seto dst

and a `Word64_addCheckP` primitive is initially translated to:

    movl src1Lo,tmpLo
    movl src1Hi,tmpHi
    addl src2Lo,tmpLo
    adcl src2Hi,tmpHi
    seto dst

which, if src1Hi were 0, can be correctly rewritten to:

    movl src1Lo,tmpLo
    xorl tmpHi,tmpHi
    addl src2Lo,tmpLo
    adcl src2Hi,tmpHi