Support OpenBSD 6.4.
OpenBSD 6.4 introduced checking that the contents of the
hardware stack pointer point into a region of memory mapped
with the `MAP_STACK` mmap(2) flag.  If, on entry to the
kernel, the %rsp on AMD64 or %esp on x86 contain other data,
the program will be killed.

Without this, executables build with mlton such as `mllex`
were dumping core.

We support this by:

1) Unconditionally reserving %rsp/%esp in builds targeting
   OpenBSD.
2) Providing plumbing support for passing arbitrary flags
   when allocating a signal stack, and passing with the
   `MAP_STACK` flag down that plumbing when defined.

With these changes, building executables for OpenBSD seems
to work as expected.

Thanks to Matthew Fluet for pointing me at the existing
support for reserving %*sp for Cygwin.

Signed-off-by: Dan Cross <crossd@gmail.com>