Fix a latent bug in SimplifyTypes related to Useless and MLton_bogus
A forthcoming SplitTypes optimization revealed a latent bug in
SimplifyTypes, related to MLton_bogus primitives added by Useless.
Unfortunately, it has been difficult constructing a simple source SML
program that demonstrates the bug.  Nonetheless, here is the essence
of the issue:

mlton.useless.pre.ssa:

    lambdas_6803 = Env_24230

    global_70971: lambdas_6803 = new Env_24230

      L_215729 (x_331531: ((word8) vector, unit) tuple)
        ...
        x_331538: (lambdas_6803, (list_161) ref) tuple =
          (global_70971 (*new Env_24230*), openInbuffers_3)
        ...
        x_331537: lambdas_28893 = new Env_44439 (x_331538)
        ...
        newIn_2: lambdas_17774 = new Env_48535 (x_331532)
        call L_215730 (newIn_0 (x_331532, global_100698)) handle _ => L_215665
      L_215730 (x_331541: instream_0)
        tuple_1834: (lambdas_17775, lambdas_4786, lambdas_17774) tuple =
          (global_70725 (*new Env_24 ... l_37)*), openf_1, newIn_2)
        ...

Useless determines that the `lambdas_6803` component of `x_331538` is
useless, but must exist.  (Must exist because somewhere else in the
program is a useful instance of a `lambdas_6803` component of a
`(lambdas_6803, (list_161) ref) tuple` and that tuple flows to a
program point with a useless instance of a `lambdas_6803` component of
a `(lambdas_6803, (list_161) ref) tuple`; in order to avoid a
coercion, Useless requires the `lambdas_6803` component to exist.)  To
fulfill the obligation for a value of a type that is useless but must
exist (and the original value is not otherwise useful and retained by
the program), Useless creates a (bogus) value with `MLton_bogus`.  In
this particular case, retaining the nullary `Env_24230` global would
suffice, but, in general, an unbounded amount of work could be saved
by using `MLton_bogus` rather than retaining the value from the input
SSA program.

Also, because the `lambdas_6803` component of `x_331538` is useless,
`global_70971` is useless and removed.

mlton.useless.post.ssa:

    lambdas_6803 = Env_24230

    bogus_79: lambdas_6803 = prim MLton_bogus(lambdas_6803) ()

      L_215729 (x_331531: (word8) vector)
        ...
        x_331538: (lambdas_6803, (list_161) ref) tuple =
          (bogus_79, openInbuffers_3)
        ...
        x_331537: lambdas_28893 = new Env_44439 (x_331538)
        newIn_2: lambdas_17774 = new Env_48535 (x_331532)
        call L_217678 (newIn_0 (x_331532, global_100698)) handle _ => L_215665
      L_217678 (x_338600: instream_0)
        tuple_1834: (lambdas_17775, lambdas_4786, lambdas_17774) tuple =
          (global_70725 (*new Env_24 ... l_37)*), openf_1, newIn_2)
        ...

This portion of the program is unchanged through to removeUnused2.

mlton.removeUnused2.pre.ssa2:

    lambdas_6803 = Env_24230

    bogus_79: lambdas_6803 = prim MLton_bogus(lambdas_6803) ()

      L_215729 (x_331531: (word8) vector)
        ...
        x_331538: (lambdas_6803,  (list_161) ref) tuple =
          (bogus_79, openInbuffers_3)
        ...
        x_331537: lambdas_28893 = new Env_44439 (x_331538)
        ...
        newIn_2: lambdas_17774 = new Env_48535 (x_331532)
        call L_217678 (newIn_0 (x_331532, global_100698)) handle _ => L_215665
      L_217678 (x_338600: instream_0)
        tuple_1834: (lambdas_17775,  lambdas_4786,  lambdas_17774) tuple =
          (global_70725 (*new Env_24 ... l_37)*), openf_1, newIn_2)
        ...

The previously eliminated `global_70971` was the only use of
`Env_24230` in the program, so `removeUnused2` eliminates it.
However, `lambdas_6803` cannot be eliminated because it appears in
types and primitives, in particular, `MLton_bogus` for `bogus_79`.

mlton.removeUnused2.post.ssa2:

    lambdas_6803 =

    bogus_79: lambdas_6803 = prim MLton_bogus(lambdas_6803) ()

      L_215729 (x_331531: (word8) vector)
        ...
        x_331538: (lambdas_6803,  (list_161) ref) tuple =
          (bogus_79, openInbuffers_3)
        ...
        x_331537: lambdas_28893 = new Env_44439 (x_331538)
        ...
        newIn_2: lambdas_17774 = new Env_48535 (x_331532)
        call L_222685 (newIn_0 (x_331532, global_100698)) handle _ => L_215665
      L_222685 (x_343607: instream_0)
        tuple_1834: (lambdas_17775,  lambdas_4786,  lambdas_17774) tuple =
          (global_70725 (*new dummy_1852*), openf_1, newIn_2)
        ...

Again, this portion of the program is unchanged into simplifyTypes.

mlton.simplifyTypes.pre.ssa:

    lambdas_6803 =

    bogus_79: lambdas_6803 = prim MLton_bogus(lambdas_6803) ()

      L_215729 (x_331531: (word8) vector)
        ...
        x_331538: (lambdas_6803,  (list_161) ref) tuple =
          (bogus_79, openInbuffers_3)
        ...
        x_331537: lambdas_28893 = new Env_44439 (x_331538)
        ...
        newIn_2: lambdas_17774 = new Env_48535 (x_331532)
        call L_222685 (newIn_0 (x_331532, global_100698)) handle _ => L_215665
      L_222685 (x_343607: instream_0)
        tuple_1834: (lambdas_17775,  lambdas_4786,  lambdas_17774) tuple =
          (global_70725 (*new dummy_1852*), openf_1, newIn_2)
        ...

Because there are no constructors for `lambdas_6803`, the cardinality
of `lambdas_6803` is determined as `Zero`.  This makes the cardinality
of `(lambdas_6803, (list_161) ref) tuple` and `Env_44439` both `Zero`.
Because the cardinality of `Env_44439` is `Zero`, it is removed from
`lambdas_17774` and marked as `Useless`.  Finally, a `ConApp` of a
`Useless` constructor is considered unreachable code (because such a
constructor must be constructed with an argument of cardinality
`Zero`, but (`MLton_bogus` not withstanding), no such value could have
been constructed by the program).  Upon encountering
`x_331537: lambdas_28893 = new Env_44439 (x_331538)`, SimplifyTypes
stops translating the statements of `L_215729` (and translates the
whole block into a statementless/Bug-transfer block).  However,
SimplifyTypes iterates through the blocks of the program via a DFS, so
goes on to translate `L_222685` and encounters a use of `newIn_2` and
terminates with `newIn_2 has no varInfo property`.

In this particular program, it is not appropriate to conclude that
`lambdas_6803` has cardinality `Zero`, because a "value" of such a
type is created by `MLton_bogus`.  To accommodate such "values", the
SimplifyTypes pre-processing is extended to mark any tycon created by
`MLton_bogus` as having cardinality `Many`.  It does not seem
necessary to mark all constituent types of the type created by a
`MLton_bogus` as having cardinality `Many`, because a "bogus" values
should never be deconstructed.

It is not clear that the treatment of unreachable code following the
construction of a `Zero` cardinality ConApp is sufficiently
robust (because of the DFS that will visit successor blocks), but has
not seemed to be problematic.