Fix latent bug in Useless with `Ref_ref` and `Weak_new` primitives
Consider the following program:

    val r1 : int ref = ref 1
    val r2 : int ref = ref 2
    val chk = r1 = r2
    val () = print (concat [Bool.toString chk, "\n"])

A forthcoming update to the Useless optimization will not force the
contents of `array`, `ref`, and `weak` types as useful when the type
participates in `MLton_equal` (because polymorphic equality does not
inspect the components of types with identity).  This will make the
contents of `r1` and `r2` useless; however, the values `1` and `2`
likely remain useful elsewhere in the program.  Hence, for `Ref_ref`
and `Weak_new` primitives, it does not suffice to map over the
arguments and retain those that are useful.

For `Ref_ref` and `Weak_new`, if a `(unit) ref` or `(unit) weak` is
being constructred, then ignore the existing arguments and build the
object with `unitVar`.