Strengthen Useless optimization wrt MLton_equal and MLton_hash
The MLton_equal (polymorphic equality) and MLton_hash (polymorphic
hashing) primitives were analyzed by applying `deepMakeUseful` to
their arguments.  However, when implemented, these primitives do not
descend into Array, Ref, or Weak types, so marking the contents of
Array, Ref, and Weak types as useful is unnecessary and misses
opportunities to eliminate useless components.

In particular, while investigating the issue described in 19b07c0, it
was observed that the tuple component that was useless but must exist
was only forced to exist due to another tuple's corresponding
component being made useful by the `deepMakeUseful` triggered by
analyzing MLton_equal.  However, this was confusing, as the tuple
component in question was a `lambdas_nnn` type, which should not be
compared by polymorphic equality.  Further investigation revealed that
it was a `MLton_equal((list_mmm) ref)`, with `list_mmm` corresponding
to a list of lambdas.  This had the effect of marking *many*
transitive closure components as useful, even though they would not
actually be used by the implementation of polymorphic equality.

This strengthening of Useless has a non-trivial effect on a
self-compile:

pre.useless:
               pre.ssa size = 146,403,800 bytes
               num globals = 37936
               num vars in main = 10709
               num blocks in main = 2082
               num functions in program = 7711
               num blocks in program = 231863
               num statements in program = 308427
               num types in program = 21671

post.useless (old):
               post.ssa size = 170,454,040 bytes
               num globals = 19772
               num vars in main = 10264
               num blocks in main = 2082
               num functions in program = 7711
               num blocks in program = 230017
               num statements in program = 283672
               num types in program = 18112

post.useless (new):
               post.ssa size = 167,373,184 bytes
               num globals = 19084
               num vars in main = 9628
               num blocks in main = 2067
               num functions in program = 7711
               num blocks in program = 229503
               num statements in program = 279612
               num types in program = 17065