Eliminate "weirdness with raise/handle" in Useless
Previously, the Useless pass would unify all handler arguments so that
raise/handle had a consistent calling convention.  This seems to be
a (long) hold-over from the CPS IL with explicit HandlerPush and
HandlerPop statements.

Now, allow functions to raise more than is used by a handler,
inserting coercion wrapper blocks as necessary.  (Something very
similar is done in RemoveUnsed.)  While this can eliminate useless
components of a function's `raises`, it can also require turning a
Return.Tail into a Return.NonTail or a Handler.Caller into a
Handler.Handle, in order to discard useless components.

On a self-compile, compared to previously:
 * 397 functions go from `raises = Some (exn)` to `raises = Some ()`
 * 3 calls go from Return.Tail to Return.NonTail
 * 1571 calls go from Return.Handler to Handler.Handle