Remove `size2` and `size3` regression tests
A forthcoming update to the `Useless` optimization to address
MLton/mlton#201 will not force the argument of `MLton_size` to be
"deeply useful".  However, nearly all of the data structures in the
`size2` and `size3` regression tests are only used for their
size (and, so, the new `Useless` optimization mostly eliminates the
argument to `MLton_size` leading to 0 sizes).  It has proven too
difficult to robustly use the values after measuring their size.
Moreover, the regression tests have been awkward, due to their
sensitivity to the architecture.