Rework `size` regression test
A forthcoming update to the `Useless` optimization to address
MLton/mlton#201 will not force the argument of `MLton_size` to be
"deeply useful".  However, many of the data structures in the `size`
regression test are only used for their size (and, so, the new
`Useless` optimization mostly eliminates the argument to `MLton_size`
leading to 0 sizes).  This commit adds an explicit "use" of the values
after measuring their size.  In addition, since the `size2` and
`size3` regression tests have been removed, the `size` regression test
now outputs exact sizes (except for the "continuation option ref") in
order to monitor changes to object sizes.