Strengthen Useless optimization wrt MLton_{eq,share,size}
Closes MLton/mlton#201

The `Useless` optimization eliminates components of of tuples and
constructors that are not observed by the computation.  Previously,
`MLton_share` was analyzed as a side-effect primitive (it must be
preserved by optimizations to perform as the programmer intended) and
side-effect primitives have `deepMakeUseful` applied to their
arguments; this effectively forced all components of the types flowing
to `MLton_share` to be retained.  Similarly, although `MLton_eq` and
`MLton_size` are not considered side-effect primitives, if the result
of `MLton_eq`/`MLton_size` was considered useful, then all components
of the types flowing to them were retained.

Now, we allow the type upon which these primitives are acting to
change as dictated by the rest of the program.  In particular, we
split the single "useless/useful" lattice element into a pair of
"useless/useful" and "not wanted/wanted" lattice elements.  Flipping
from "not wanted" to "wanted" causes any "not wanted/wanted" element
flowing into the current element to be flipped to "wanted" and unifies
with any "useless/useful" element flowing into the current element.
This has the effect that a component that is "wanted" is also "useful"
if any "useful" component flows to it.