Rework `weak.2` regression test
A forthcoming update to the `Useless` optimization will not force the
argument to `Weak_canGet` to be "deeply useful".  However, the tuple
used in the `weak.2` regression test is not actually used (and, so,
the new `Useless` optimization changes the `(int * int ref) Weak.t` to
a `unit Weak.t`, which is considered never alive).  This commit adds
an explicit "use" of the tuple.