Strengthen Useless optimization wrt Weak_canGet
The `Useless` optimization eliminates components of of tuples and
constructors that are not observed by the computation.  Previously,
`Weak_canGet` was analyzed as a non-side-effect primitive; if the
result of `Weak_canGet` was considered useful, then all components of
the argument of `Weak_canGet` were considered useful.

However, 'Weak_canGet` does not actually use the components of the
weak reference; rather, it only uses the reference itself.