Strengthen Useless optimization wrt MLton_touch
The `Useless` optimization eliminates components of data constructors
that are not observed by the computation.  Previously, `MLton_touch`
was analyzed as a side-effect primitive; side-effect primitives have
`deepMakeUseful` applied to their arguments; this effectively forced
all components of the types flowing to `MLton_touch` to be retained.
However, 'MLton_touch` does not actually use the data structure; it
simply keeps the data structure alive.  The typical usage of
`MLton_touch` for finalizers is to `touch` a reference; hence, it
suffices to use `shallowMakeUseful` to make the reference useful,
without necessarily making the contents deeply useful.