Perform elimination of common globals in CommonSubexp
Various passes may introduce redundant globals.  The DuplicateGlobals pass
specifically introduces redundant globals (that may be separated by SplitTypes),
which should be recombined.  But, various other passes introduce `true`,
`false`, `0`, `1`, etc. global constants in order to avoid needing to search the
globals for a particular constant (that is likely, but not guaranteed, to
exist).