Implement `ID.parse: t Parse.t` and related functions
Parsing of identifiers is complicated, even when restricted to
well-formed IL programs with single defs of each variable.

Each def and use of the same variable in a printed IL program must be
mapped to the same ID object.  To handle this, a hash table mapping
strings to ID objects is maintained.

Some IDs are special (e.g., the Tycon corresponding to `bool`),
because there is a particular (global) ID object for that identifier.
To handle this, `val parseReset: {prims: t vector} -> unit` is used to
remove all entries from the hash table and insert a number of
primitive identifiers into the hash table (according to the
identifier's `toString`).

Non-special IDs are printed with a uniquifying suffix (e.g., `x_123`),
but if that identifier is parsed and then printed again, we do not
wish to add another uniquifying suffix identifier is printed (e.g.,
`x_123_456`).  To handle this, parsing strips the uniquifying suffix
when creating the ID object (using `newString`).  A more sophisticated
implementation would strip the uniquifying suffix, create the ID
object with an originalName and a printName, and update the
UniqueString table based on the suffix.

Finally, it is sometimes useful to handle tokens that could appear
where an identifier could occur, but should not be taken as an
identifier.  (For example, in the SSA2 IL, an ObjectCon is either a
Con identifier, "tuple", or "sequence".)  To handle this,
`val parseAs: (string * 'a) vector * (t -> 'a) -> 'a Parse.t` can be
used.  An alternate design would be
`val parseAs: (string -> 'a option) * (t -> 'a) -> 'a Parse.t`.