Implement `XmlTree.Program.parse`
Reimplement the XML IL parser in `functor XmlTree`.  This
implementation places the parser for each IL component next to the
layout for the corresponding component.  This will make it much easer
to keep the layout and parser for each component in sync.

This implementation uses a rather more monadic than applicative style
of parser, which may be less efficient, but seems to be more
maintainable (at least with the present set of parser combinators).