Fix SSA2 to RSSA translation of GC_collect primitive
The GC_collect primitive is translated to a C call to GC_collect,
which has `ensuresBytesFree = true`.  However, no `EnsuresBytesFree`
operand is created for the C call; instead, a constant 0 operand was
being used for the `bytesRequested` argument.  During the
`insertLimitChecks` pass, a C call with `ensuresBytesFree = true` is
expected to have an `EnsuresBytesFree` operand that can be replaced by
the amount of heap space required by the C call return block.  Failing
to pass the appropriate `bytesRequested` allows the garbage collection
to shrink the heap and fail to ensure that the necessary heap space is
available.