Simplify C and LLVM codegen trampoline
Previously, a chunk had the prototype `struct cont ChunkN(void)`, a
global `uintptr_t nextFun` contained the index of the next block to
execute, while a `structure cont { void* nextChunk; }` contained the
next chunk to execute, so the trampoline looked like:

   cont=(*(struct cont(*)(void))cont.nextChunk)();

Returning the next block index via a global seems inefficient (and
does not work in the CMU MultiMLton project).

Now, a chunk has the prototype `uintptr_t ChunkN(uinptr_t)`, where the
argument is the index of the next block to execute and the result is
the index of the next block to execute, so the trampoline looks like:

   nextChunk = nextChunks[nextBlock];
   nextBlock = (*nextChunk)(nextBlock);

In addition, the global `int returnToC`, used to signal the
termination of a call to an exported SML function, is eliminated.
Instead, the special block index `(uintptr_t)-1` is used to terminate
the trampoline.

One minor inefficiency is that the known destination chunk of a call
cannot be exploited.