Make `CReturn<ty>` variables chunk local
Previously, the `CReturn<ty>` variables were globals, declared in the
"main" .c file and `extern` imported into the chunk .c and .ll files.
This is unnecessary (and does not work in the CMU MultiMLton project);
in particular, it requires the c compiler to write the result to
memory and, depending on the intervening code, to read the result from
memory, although the result value is likely to have been returned via
a register.

Now, the `CReturn<ty>` variables are chunk local.  They are marked
`__attribute__ ((unused))` to suppress compiler warnings.  Although it
would be possible to examine the set of CCalls in the chunk (as is
done to declare the FFI calls) and only generate variables that
correspond to return types of called functions, even this does not
suffice, because a CReturn block can ignore the result and the
`CReturn<ty>` variable is unused.