Split CFunction.Kind.Runtime maySwitchThreads field into maySwitchThreads{From,To}
maySwitchThreadsFrom indicates that the C function may switch ML
threads; at the return from the C function, the mutator should jump to
the block indicated by the return address on the top of the now
current ML stack.  Examples are Thread_switchTo and GC_collect (with
signals).

maySwitchThreadsTo indicates that the C function's return block may be
switched to, in the sense that that block may be indicated by the
return address on the top of a stack seen by a `maySwitchThreadsFrom`
function.  Examples are Thread_switchTo, GC_collect (with signals),
GC_copyCurrent, and GC_saveWorld; note that these last two functions
will always return and resume execution with the C function's return
block (the ML thread is not switched by these functions), but explicit
thread switching or world resumption will also begin execution at the
C functions' return block.

Note that maySwitchThreadsFrom implies maySwitchThreadsTo.

This change is in preparation for a change to the C and LLVM codegens
to treat `maySwitchThreadsTo` as the indicator of an entry, rather
than `mayGC`.