Require `maySwitchThreadsTo`, rather than `mayGC`, for CReturn entry
Treating a block as an entry requires that the block have an index in
the `ChunkSwitch`.  However, it is unnecessary for `mayGC=true` but
`maySwitchThreadsTo=false` to be entries.  Furthermore, it can lead to
C compiler warnings, due to the following:

    ChunkSwitch (0)
    ...
    case 5: goto L_45;
    ...
    EndChunkSwitch

    L_44:
            S(Word64, 56) = /* L_45 */ 5;
            Push (64);
            FlushFrontier ();
            FlushStackTop ();
            CReturnP = GC_sequenceAllocate (GCState, (Word64)(0x0ull), (Word64)(0x21ull), (Word64)(0x1Bull));
            CacheFrontier ();
            CacheStackTop ();
            goto L_45;

    L_45:
            Push (-64);
            S(Objptr, 56) = CReturnP;
            W64_0 = (Word64)(0x0ull);
            goto loop_6;

`GC_arrayAllocate` has `mayGC=true` and `maySwitchThreadsTo=false`.
Making `L_45` an entry block makes it appear that control can enter
the chunk and transfer to `L_45`, without assigning `CReturnP`.

Now, `L_45` is not treated as an entry block and the only path to
`L_45` is from the `GC_arrayAllocate` call.