Treat `MLton_halt` as a normal (non-runtime) primitive
Don't translate `MLton_halt` to a SSA `Runtime` transfer; the only use
of `Runtime` transfer is to identify `Thread_copyCurrent`.  Also,
don't translate `MLton_halt` to an RSSA `CCall` with `return=NONE`.
Omitting the return label does not push a return address to the ML
stack, which does not allow `MLton_halt` to trace the stack.