Allocate a "frame" for `Func` blocks in Machine IR
Frames serve a number of distinct purposes.  One purpose is to serve
as the unique index for "entry" blocks in the C and LLVM codegens.
`Func` blocks are "entry" blocks and, previously, were assigned
indices after all of the frame indices of existing blocks.  This has
the disadvantage of making the `Func` entry indices disjoint from the
other block indices, possibly inhibiting turning the chunk entry
switch into a jump table.  Moreover, with 6bc091d, the set of entry
blocks for a chunk was reduced for `CReturn` blocks from those that
have `mayGC=true` to those that have `maySwitchThreadsTo=true`.  This
further fragments the chunk entry switch.

To facilitate making the chunk entry switch a dense switch, `Func`
blocks are assigned frames.  This frame is never used by the runtime
system, since a `Func` label is never the stack top.  (Note: `Handler`
blocks are also assigned frames, and this frame is also never used by
the runtime system, since a `Handler` label is never the stack top.
On the other hand, the "size" of a `Handler` frame is used to pop the
stack.)  Hence, the frame is assigned the empty list of live offsets
and a size of 0.  This means that, for the native codegen, all `Func`
blocks will share the same frame layout and we only "waste" one entry
in the `frameLayouts` array (and, potentially, one entry in the
`frameOffsets` array, if the empty list of offsets is not otherwise
created).  For the C and LLVM codegens, we will create a distinct
entry in the `frameLayouts` array for each `Func` block; since there
are an order of magnitude more `Cont`, `Handler`, and `CReturn` blocks
than `Func` blocks, this should not be significant in code size.