Reindex frame infos for C and LLVM codegens
Adjust the frameInfo indices to ensure that the entries of each chunk
are consecutive and all entry indices precede all non-entry indices.
The former ensures that the C compiler can use a jump table for the
chunk switch.  The latter will allow the `nextChunks` array to avoid
including non-entry entries and, consequently, be much smaller.