Experiment passing stackTop and frontier as chunk arguments
With `-chunk-tail-call true`, this avoids writing and reading of
`stackTop` and `frontier` to/from the global `gcState`, because
`stackTop` and `frontier` will typically be registers within a chunk
and will typically be passed in registers according to the calling
convention.

With `-chunk-tail-call false`, the reading of `stackTop` and
`frontier` from the global `gcState` occurs in the trampoline.