Introduce RSSAShuffle pass, disabled by default
A new RSSAShuffle pass can be run after simplifying the RSSA IR
program.  This pass simply shuffles the list of functions within the
program and shuffles the vector of blocks within a function.

The motivation of RSSAShuffle is that `-chunkify coalesce<n>` has
a (small) dependency on the order of functions and blocks.