Simplify code to generate prologue of `Cont` and `Handler` blocks
f2b08b921 (Omit copying unused args of Cont and Handler blocks from
stack) introduced code to check whether or not a `Cont` or `Handler`
block argument was used by inspecting the `live` set.  Then
136f3573c (Omit allocating registers for unused args of Cont, CReturn,
and Handler blocks) introduced code to check whether or not a `Cont`
or `Handler` block argument was assigned a location.

However, `Cont` or `Handler` block argument is not assigned a location
exactly when it is not a member of the `live` set.  Hence, the
original code to check whether or not a `Cont` or `Handler` block
argument was used by inspecting the `live` set is redundant; the code
to check whether or not a `Cont` or `Handler` block argument was
assigned a location suffices.