Add `sourceSeqIndex` to `Machine.FrameInfo.T`/`struct GC_frameInfo`
By including the `sourceSeqIndex` in the frame info, frames that share their
sourceSeqIndex and all other frame info fields are automatically shared.  With
defunctorization and monomorphisation and inlining, it is quite possible for
distinct blocks of the final RSSA IR program to nontheless correspond to the
same source location.  This avoids unnecessarily creating extraneous frame infos
and simplifies the logic of when to create unique frame infos.