Eliminate `ControlFlags.elaborateOnly`
The `elaborateOnly` flag was initialized by

    stop = Place.TypeCheck
    andalso not (warnMatch)
    andalso not (!keepDefUse)

However, in order for `warnMatch` to be `false`, all of the match
warnings must be ignored and disabled, which is not typical.
Similarly, in order for `!keepDefUse` to be `false`, the `warnUnused`
annotation must be disabled, which is not typical.

It seems that there is little to be gained by trying to optimize the
`elaborateOnly` path.