Remove Parse.compose
There are a couple issues with the existing Parse.compose:
1. It acts incorrectly if the second parser fails to consume all input
from the first: The remainder of the output from parser 1 will be left
in the state, and be processed again. This is unproblematic in our
particular usage: removing comments at the top-level of a program parse.
2. It is very difficult to implement efficiently. Since the parsers
now work on vectors, it seems difficult to efficiently produce the
vectors desired by the second parser.

Since the only usage present was by comment removal, and this
only produced a space, single character, or string (that may contain
comment delimeters), it will be integrated into other methods
consuming spaces or strings, which is also a simplification.

This means that some parse errors are different: namely,
data(* a comment *)type will give an error upon reaching the first
'(', whereas compose would have produced 'data type', which
could give an error at the space or the start of type.