Introduce `GC_state` primitive, translated to `GCState` RSSA operand
This primitive will replace the use of `_symbol "gcStateAddress"` in
the Basis Library implementation.

One motivation is that it eliminates two "names" for the same thing;
the RSSA `GCState` operand and the pointer fetched from
`gcStateAddress`.

Another motivation is that the CMU MultiMLton compiler cannot use
`_symbol "gcStateAddress"`, because this fetches the primary (aka,
pthread0) `gcState` address once at the beginning of the evaluation of
the Basis Library, but each use of the value should actually be with
respect to the `gcState` of the executing pthread.

In the Basis Library, the type of `Primitive.MLton.GCState.t` is
`Pointer.t` (which is the primitive type `cpointer`, which becomes the RSSA IL
type `CPointer`), but in the RSSA IL, the type of the `GCState` operand is
`GCState`.  One option would be to introduce a new primitive type `gcstate` to
become the RSSA `GCState` type; however, this would affect all of the
intermediate ILs.  Moreover, the RSSA type checking of `GCState` is somewhat
weak, so it seems simpler to simply replace the RSSA `GCState` type with the
RSSA `CPointer` type.  (Similarly, we replace the RSSA `ExnStack` type with the
RSSA `Word` type of the appropriate size.)