Use `GC_state` primitive instead of `_symbol "gcStateAddress"`
One motivation is that it eliminates two "names" for the same thing;
the RSSA `GCState` operand and the pointer fetched from
`gcStateAddress`.

Another motivation is that the CMU MultiMLton compiler cannot use
`_symbol "gcStateAddress"`, because this fetches the primary (aka,
pthread0) `gcState` address once at the beginning of the evaluation of
the Basis Library, but each use of the value should actually be with
respect to the `gcState` of the executing pthread.