Raise values via stack with `-raise-style stack`
With `-raise-style globals`, raised arguments are communicated from
the raiser to the handler via (non-root) globals.  This leads to the
confusing `globalObjptrNonRoot`.  Moreover, it has been a source of
bugs in the CMU MultiMLton project, where parallel threads that raise
at the same time can interfere through the race on the handler
globals.

With the new `-raise-style stack`, raised arguments are communicated
from the raiser to the handler via the ML stack.  Sufficient space for
raised values are reserved in the handler frame immediately above the
handler label.  At a raise, the raiser stores the raised values at
offsets of `stackBottom + exnStack` and the handler copies the values
from the adjusted offsets to its argument locations.  (Thus, a
`Handler` block executes in the same manner as a `Cont` block.)

The downside of this strategy is that every stack frame with an
installed handler requires an additional 8 bytes (on a 64-bit
platform) for the raised value, which is above the 16 bytes for the
handler label and the exception stack link.  (In theory, more space
could be required if earlier optimizations flattened the `raises` of
an SSA/SSA2/RSSA function, which could happen with the `SplitTypes`
optimization that separates distinct "uses" of exceptions.  However,
in practice, the only "raise" convention has been a single `objptr`.)

A different strategy would be for the raiser to store the raised
values in its frame and then communicate the address of the raiser's
frame to the handler.  However, this would require an additional
global to communicate the address of the raiser's frame.