Add lastError and stack to parse state
lastError allows us to possibly give better messages
when we have a case like `many x *> y`, so that the error
reported comes from trying to parse x rather than y
(which we expect to fail more easily). It is not
fully functional in this commit.

Putting the stack in the parse state should improve
the performance of adding to the stack versus stacking
closures