Change AllocateRegisters to AllocateVariables
Since there are an infinite number of temporaries and stack slots,
and we use only a simple greedy algorithm for this,
the name AllocateRegisters (resp. AllocateTemporaries) is a bit
of a misnomer. Rather than assign each temporary or stack slot
to a variable, we assign to each variable a stack slot or temporary
with no limit to the amount available.