Experiment with forcing jump table for C codegen
Using GCC's label address and computed goto
features (https://gcc.gnu.org/onlinedocs/gcc/Labels-as-Values.html#Labels-as-Values),
one can force the generation of a jump table for the chunk switch.

Although GCC and clang will typically implement dense `switch`
statements into a jump table, consider the following:

    switch (nextBlock) {
      case 5: goto L_5;
      case 6: goto L_6;
      ...
      case 29: goto L_29;
      default: __builtin_unreachable();
    }

GCC-7 (and earlier) and clang appear to implement this as:

   int t = nextBlock - 5
   if (t > 24)
     goto L_29;
   else
     goto *jumpTable[i];

That is, it still performs a range comparison.  With an explicit jump
table, GCC and clang will implement this as:

   goto *jumpTable[nextBlock - 5];

(where the -5 can be incorporated into the address computation).

Unfortunately, the performance impact seems negligible.